requirements.txt
prettyprinter~=0.18.0
attrs~=21.2.0
setuptools~=56.0.0
pytz~=2021.1
python-dateutil~=2.8.1
aiohttp~=3.7.4